An explanation of the comment bugs:
The previous implementation of comment handling by JavaCC functioned as
follows:-

When seeing “;;” - enter a lexical state of IN_COMMENT, with the comment
terminated by a line-end.

This had two inherent problems:-

1.	In the GUI, the line end was really a virtual enter key (VK_ENTER)
event – it was never present in the input stream and therefore the
IN_COMMENT state was never left. The solution is to append the system
dependant line-end to the input when appropriate
2.	In both the shell and the GUI (once the above was resolved) the
determination of the IN_COMMENT state was dependant on whether the line
was being wholly processed within the parser or was being processed by
application code. The following held true

;; This is a working comment
(deftemplate testtemplate
            ;; this is a working comment
            (slot test1) ;; this is also a working comment
            (slot ;; this comment would not work, but why would you do
this?
                        test2)
            )

;; Another working comment

(defrule testrule
            (testtemplate (test1 ?myval))
=>
            ;; This comment will eat the following line
            (printout t ?myval crlf)
            )

The parser did not detect the end of the comment until the end of the
following line

There are several solutions (looking ahead, etc) but the simplest and
most robust seems to be to just thrown away any text starting ;; and
ending in a line-end